Daniil feature 4

+++ app/Http/Controllers/RoleController.php
@@ -6,14 +6,21 @@ use Illuminate\Http\Request;
 use App\Http\Requests\Role\CreateRoleRequest;
 use Illuminate\Support\Facades\DB;
 use App\Models\Role;
+use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
     public function index()
     {
+        if (Gate::denies('view', User::class)) {
+            throw new AuthorizationException('У вас нет разрешения на просмотр прав.');
+        }
+
         $roles = Role::query()
-        ->latest()
-        ->get();
+            ->latest()
+            ->get();
         return view('roles.index', compact('roles'));
     }
 
@@ -32,12 +39,16 @@ class RoleController extends Controller
     public function show(Role $role)
     {
         $permissions = $role->permissions()->get();
-        
+
         return view('roles.show', compact('role', 'permissions'));
     }
 
     public function delete(Role $role)
     {
+        if (Gate::denies('delete', User::class)) {
+            throw new AuthorizationException('У вас нет разрешения на просмотр прав.');
+        }
+
         DB::transaction(function () use ($role) {
             $role->permissions()->detach();
             $role->users()->detach();
